// src/app/api/roboticsAudit/audit.ts
export interface audit {
    audit_id?: number;
    audit_uid?: string;
    audit_board_id?: string;
    audit_name?: string;
    banner_id?: number;
    au_level_2_id?: number;
    review_year?: number;
    quarter?: number;
    start_date?: string;
    end_date?: string;
    acp_audit?: boolean;
    user_access_id?: string;
    audit_schedule?: string;
    results_url?: string;
    last_run_date?: string;
    banner?: string;
    banner_uid?: string;
    au_level_2_uid?: string;
    au_level_2_desc?: string;
    full_name?: string;
    useracsname?: string;
}